Adding test rewards action
