+++ .github/action_scripts/check_if_node_started/index.js
