+++ .github/templates/dag_l1/initial_validator/action.yml
@@ -61,4 +61,4 @@ runs:
 
         nohup java -jar ../../jars/dag-l1.jar run-initial-validator --ip 127.0.0.1 > dag-l1-initial-validator.log 2>&1 &
 
-        node ../../../../action_scripts/check_if_node_started.js -url=http://127.0.0.1:${{ inputs.CL_PUBLIC_HTTP_PORT }}/node/info -cluster_name=DAG-L1-Initial-Validator
+        node ../../../../action_scripts/check_if_node_started/index.js -url=http://127.0.0.1:${{ inputs.CL_PUBLIC_HTTP_PORT }}/node/info -cluster_name=DAG-L1-Initial-Validator
