+++ .github/templates/dag_l1/validator/action.yml
@@ -52,7 +52,7 @@ runs:
 
         nohup java -jar ../../jars/dag-l1.jar run-validator --ip 127.0.0.1 > dag-l1-validator-${{ inputs.NODE_NUMBER }}.log 2>&1 &
 
-        if node ../../../../action_scripts/check_if_node_started.js -url=http://127.0.0.1:${{ inputs.CL_PUBLIC_HTTP_PORT }}/node/info -cluster_name=DAG-L1-Validator-${{ inputs.NODE_NUMBER }}; then
+        if node ../../../../action_scripts/check_if_node_started/index.js -url=http://127.0.0.1:${{ inputs.CL_PUBLIC_HTTP_PORT }}/node/info -cluster_name=DAG-L1-Validator-${{ inputs.NODE_NUMBER }}; then
           sleep 5
           echo "Joining dag-l1 validator node ${{ inputs.NODE_NUMBER }} ..."
           curl -X POST http://127.0.0.1:${{ inputs.CL_CLI_HTTP_PORT }}/cluster/join -H "Content-type: application/json" -d "{ \"id\":\"$DAG_L1_INITIAL_VALIDATOR_NODE_ID\", \"ip\": \"127.0.0.1\", \"p2pPort\": ${{ inputs.CL_L1_INITIAL_VALIDATOR_P2P_PORT }} }"
