+++ .github/templates/metagraph_l0/ml0_cluster/action.yml
@@ -1,5 +1,9 @@
 name: Metagraph L0 Cluster
 
+inputs:
+  IMPLEMENT_REWARDS:
+    required: false
+
 runs:
   using: "composite"
   steps:
@@ -8,6 +12,7 @@ runs:
         CL_PUBLIC_HTTP_PORT: 9400
         CL_P2P_HTTP_PORT: 9401
         CL_CLI_HTTP_PORT: 9402
+        IMPLEMENT_REWARDS: ${{ inputs.IMPLEMENT_REWARDS }}
       uses: "./.github/templates/metagraph_l0/genesis"
 
     - name: Get the Metagraph ID
