+++ .github/templates/metagraph_l0/validator/action.yml
@@ -56,7 +56,7 @@ runs:
         
         nohup java -jar ../../jars/metagraph-l0.jar run-validator --ip 127.0.0.1 > metagraph-l0-validator-${{ inputs.NODE_NUMBER }}.log 2>&1 &
         
-        if node ../../../../action_scripts/check_if_node_started.js -url=http://127.0.0.1:${{ inputs.CL_PUBLIC_HTTP_PORT }}/node/info -cluster_name=Metagraph-L0-Validator-${{ inputs.NODE_NUMBER }}; then
+        if node ../../../../action_scripts/check_if_node_started/index.js -url=http://127.0.0.1:${{ inputs.CL_PUBLIC_HTTP_PORT }}/node/info -cluster_name=Metagraph-L0-Validator-${{ inputs.NODE_NUMBER }}; then
             sleep 5
             echo "Joining Metagraph l0 validator node ${{ inputs.NODE_NUMBER }} ..."
             curl -X POST http://127.0.0.1:${{ inputs.CL_CLI_HTTP_PORT }}/cluster/join -H "Content-type: application/json" -d "{ \"id\":\"$CURRENCY_L0_GENESIS_NODE_ID\", \"ip\": \"127.0.0.1\", \"p2pPort\": ${{ inputs.CL_L0_P2P_HTTP_PORT }} }"
