+++ .github/templates/metagraph_l1/initial_validator/action.yml
@@ -29,9 +29,16 @@ runs:
     - name: Generate Metagraph L1 JAR
       shell: bash
       run: |
-        sbt currencyL1/assembly
+        cd .github/config/containers/
+        [ ! -d "github-action-metagraph-test" ] && git clone https://github.com/Constellation-Labs/github-action-metagraph-test.git
+        cd github-action-metagraph-test/
         
-        mv modules/currency-l1/target/scala-2.13/tessellation-currency-l1-assembly-* .github/config/containers/jars/metagraph-l1.jar
+        sbt currencyL1/assembly
+
+    - name: Move Metagraph L1 JAR
+      shell: bash
+      run: |
+        mv .github/config/containers/github-action-metagraph-test/modules/l1/target/scala-2.13/custom-project-currency-l1-assembly-* .github/config/containers/jars/metagraph-l1.jar
 
     - name: Running Metagraph L1 Initial Validator
       shell: bash
@@ -76,4 +83,4 @@ runs:
 
         nohup java -jar ../../jars/metagraph-l1.jar run-initial-validator --ip 127.0.0.1 > metagraph-l1-initial-validator.log 2>&1 &
 
-        node ../../../../action_scripts/check_if_node_started.js -url=http://127.0.0.1:${{ inputs.CL_PUBLIC_HTTP_PORT }}/node/info -cluster_name=Metagraph-L1-Initial-Validator
+        node ../../../../action_scripts/check_if_node_started/index.js -url=http://127.0.0.1:${{ inputs.CL_PUBLIC_HTTP_PORT }}/node/info -cluster_name=Metagraph-L1-Initial-Validator
