+++ .github/workflows/test-metagraph-rewards.yml
@@ -1,4 +1,4 @@
-name: Test sending transactions
+name: Test metagraph rewards
 
 on:
   pull_request: {}
@@ -10,7 +10,7 @@ on:
 
 jobs:
   build:
-    name: Test sending transactions
+    name: Test rewards
     runs-on: ubuntu-20.04
     timeout-minutes: 45
     strategy:
@@ -49,14 +49,11 @@ jobs:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
         uses: "./.github/templates/global_l0/gl0_cluster"
 
-      - name: Run DAG L1 Cluster
-        env:
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-        uses: "./.github/templates/dag_l1/dl1_cluster"
-
       - name: Run Metagraph L0 Cluster
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+        with:
+          IMPLEMENT_REWARDS: true
         uses: "./.github/templates/metagraph_l0/ml0_cluster"
 
       - name: Run Metagraph L1 Cluster
@@ -64,88 +61,71 @@ jobs:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
         uses: "./.github/templates/metagraph_l1/ml1_cluster"
 
-      - name: Check DAG cluster
-        run: |
-          cd .github/action_scripts/check_clusters
-          node dag.js
-
       - name: Check Metagraph cluster
         run: |
           cd .github/action_scripts/check_clusters
           node metagraph.js
 
-      - name: Send transactions
+      - name: Check rewards
         run: |
-          cd .github/action_scripts
-          npm install @stardust-collective/dag4
-          node send_transactions/index.js
+          cd .github/action_scripts/rewards
+          node index.js
+
+      - name: Wait 1 minute before check cluster
+        run: |
+          sleep 60
+
+      ## Checking cluster again because we already faced scenarios where the cluster dies after sending rewards transactions
+      - name: Check Metagraph cluster
+        run: |
+          cd .github/action_scripts/check_clusters
+          node metagraph.js
 
       - name: Save Global L0 logs
         uses: actions/upload-artifact@v3
         if: always()
         with:
-          name: send-transactions-global-l0-java-${{ matrix.java }}
+          name: rewards-global-l0-java-${{ matrix.java }}
           path: .github/config/containers/global-l0/global-l0.log
-
-      - name: Save DAG L1 Initial Validator logs
-        uses: actions/upload-artifact@v3
-        if: always()
-        with:
-          name: send-transactions-dag-l1-java-${{ matrix.java }}
-          path: .github/config/containers/dag-l1/initial-validator/dag-l1-initial-validator.log
-
-      - name: Save DAG L1 Validator - 1 logs
-        uses: actions/upload-artifact@v3
-        if: always()
-        with:
-          name: send-transactions-dag-l1-java-${{ matrix.java }}
-          path: .github/config/containers/dag-l1/validator-1/dag-l1-validator-1.log
-
-      - name: Save DAG L1 Initial Validator - 2 logs
-        uses: actions/upload-artifact@v3
-        if: always()
-        with:
-          name: send-transactions-dag-l1-java-${{ matrix.java }}
-          path: .github/config/containers/dag-l1/validator-2/dag-l1-validator-2.log
 
       - name: Save Metagraph L0 - Genesis logs
         uses: actions/upload-artifact@v3
         if: always()
         with:
-          name: send-transactions-metagraph-l0-genesis-java-${{ matrix.java }}
+          name: rewards-metagraph-l0-genesis-java-${{ matrix.java }}
           path: .github/config/containers/metagraph-l0/genesis/metagraph-l0-genesis.log
 
       - name: Save Metagraph L0 - Validator 1 logs
         uses: actions/upload-artifact@v3
         if: always()
         with:
-          name: send-transactions-metagraph-l0-validator-1-java-${{ matrix.java }}
+          name: rewards-metagraph-l0-validator-1-java-${{ matrix.java }}
           path: .github/config/containers/metagraph-l0/validator-1/metagraph-l0-validator-1.log
 
       - name: Save Metagraph L0 - Validator 2 logs
         uses: actions/upload-artifact@v3
         if: always()
         with:
-          name: send-transactions-metagraph-l0-validator-2-java-${{ matrix.java }}
+          name: rewards-metagraph-l0-validator-2-java-${{ matrix.java }}
           path: .github/config/containers/metagraph-l0/validator-2/metagraph-l0-validator-2.log
 
       - name: Save Metagraph L1 - Initial Validator logs
         uses: actions/upload-artifact@v3
         if: always()
         with:
-          name: send-transactions-metagraph-l1-initial-validator-java-${{ matrix.java }}
+          name: rewards-metagraph-l1-initial-validator-java-${{ matrix.java }}
           path: .github/config/containers/metagraph-l1/initial-validator/metagraph-l1-initial-validator.log
 
       - name: Save Metagraph L1 - Validator 1 logs
         uses: actions/upload-artifact@v3
         if: always()
         with:
-          name: send-transactions-metagraph-l1-validator-1-java-${{ matrix.java }}
+          name: rewards-metagraph-l1-validator-1-java-${{ matrix.java }}
           path: .github/config/containers/metagraph-l1/validator-1/metagraph-l1-1.log
 
       - name: Save Metagraph L1 - Validator 2 logs
         uses: actions/upload-artifact@v3
         if: always()
         with:
-          name: send-transactions-metagraph-l1-validator-2-java-${{ matrix.java }}
+          name: rewards-metagraph-l1-validator-2-java-${{ matrix.java }}
           path: .github/config/containers/metagraph-l1/validator-2/metagraph-l1-2.log
